requirements.txt
pydantic
instructor[google-generativeai]
python-dotenv